+++ src/collisions/GJK.cpp
@@ -6,10 +6,10 @@
 #include "Logger.h"
 
 
-glm::dvec3 supportFunction(const std::vector<glm::dvec3>& shape, glm::dvec3 direction) {
+glm::vec3 supportFunction(const std::vector<glm::vec3>& shape, glm::vec3 direction) {
 
 	float supportProduct = std::numeric_limits<float>::lowest();
-	glm::dvec3 supportVertex(0.0f);
+	glm::vec3 supportVertex(0.0f);
 
 	for (auto vertex : shape)
 	{
@@ -22,45 +22,45 @@ glm::dvec3 supportFunction(const std::vector<glm::dvec3>& shape, glm::dvec3 dire
 	return supportVertex;
 }
 
-glm::dvec3 minkowskiPoint(const std::vector<glm::dvec3>& shape1, const std::vector<glm::dvec3>& shape2,
-						glm::dvec3 direction) {
-	glm::dvec3 minkowski1 = supportFunction(shape1, direction);
-	glm::dvec3 minkowski2 = supportFunction(shape2, -direction);
+glm::vec3 minkowskiPoint(const std::vector<glm::vec3>& shape1, const std::vector<glm::vec3>& shape2,
+						glm::vec3 direction) {
+	glm::vec3 minkowski1 = supportFunction(shape1, direction);
+	glm::vec3 minkowski2 = supportFunction(shape2, -direction);
 	return minkowski1 - minkowski2;
 }
 
-bool pointPassedOrigin(glm::dvec3 direction, glm::dvec3 supportVertex){
+bool pointPassedOrigin(glm::vec3 direction, glm::vec3 supportVertex){
 	return dot(supportVertex, direction) > 0.0f;
 }
 
-bool createLine(std::vector<glm::dvec3>& simplex, const std::vector<glm::dvec3>& shape1,
-				const std::vector<glm::dvec3>& shape2){
-	glm::dvec3 direction = normalize(-simplex.back());
+bool createLine(std::vector<glm::vec3>& simplex, const std::vector<glm::vec3>& shape1,
+				const std::vector<glm::vec3>& shape2){
+	glm::vec3 direction = normalize(-simplex.back());
 	simplex.push_back(minkowskiPoint(shape1, shape2, direction));
 	return pointPassedOrigin(direction, simplex.back());
 }
 
-bool createTriangle(std::vector<glm::dvec3>& simplex, const std::vector<glm::dvec3>& shape1,
-					const std::vector<glm::dvec3>& shape2){
-	glm::dvec3 A = simplex[0];
-	glm::dvec3 B = simplex[1];
-	glm::dvec3 AB = B - A;
-	glm::dvec3 A0 = -A;
-	glm::dvec3 direction = normalize(cross(cross(AB, A0), AB));
+bool createTriangle(std::vector<glm::vec3>& simplex, const std::vector<glm::vec3>& shape1,
+					const std::vector<glm::vec3>& shape2){
+	glm::vec3 A = simplex[0];
+	glm::vec3 B = simplex[1];
+	glm::vec3 AB = B - A;
+	glm::vec3 A0 = -A;
+	glm::vec3 direction = normalize(cross(cross(AB, A0), AB));
 	simplex.push_back(minkowskiPoint(shape1, shape2, direction));
 	return pointPassedOrigin(direction, simplex.back());
 }
 
-bool createTetrahedron(std::vector<glm::dvec3>& simplex, const std::vector<glm::dvec3>& shape1,
-						const std::vector<glm::dvec3>& shape2)
+bool createTetrahedron(std::vector<glm::vec3>& simplex, const std::vector<glm::vec3>& shape1,
+						const std::vector<glm::vec3>& shape2)
 {
-	glm::dvec3 A = simplex[0];
-	glm::dvec3 B = simplex[1];
-	glm::dvec3 C = simplex[2];
-	glm::dvec3 BA = B - A;
-	glm::dvec3 CA = C - A;
-	glm::dvec3 A0 = -A;
-	glm::dvec3 direction = normalize(cross(BA, CA));
+	glm::vec3 A = simplex[0];
+	glm::vec3 B = simplex[1];
+	glm::vec3 C = simplex[2];
+	glm::vec3 BA = B - A;
+	glm::vec3 CA = C - A;
+	glm::vec3 A0 = -A;
+	glm::vec3 direction = normalize(cross(BA, CA));
 	if (dot(direction, A0) < 0.0f)
 	{
 		direction *= -1.0f;
@@ -69,17 +69,17 @@ bool createTetrahedron(std::vector<glm::dvec3>& simplex, const std::vector<glm::
 	return pointPassedOrigin(direction, simplex.back());
 }
 
-bool containsOriginTetrahedron(std::vector<glm::dvec3>& simplex){
-	glm::dvec3 A = simplex[0];
-	glm::dvec3 B = simplex[1];
-	glm::dvec3 C = simplex[2];
-	glm::dvec3 D = simplex[3];
-	glm::dvec3 AD = D - A;
-	glm::dvec3 BD = D - B;
-	glm::dvec3 CD = D - C;
-	glm::dvec3 DO = -D;
+bool containsOriginTetrahedron(std::vector<glm::vec3>& simplex){
+	glm::vec3 A = simplex[0];
+	glm::vec3 B = simplex[1];
+	glm::vec3 C = simplex[2];
+	glm::vec3 D = simplex[3];
+	glm::vec3 AD = D - A;
+	glm::vec3 BD = D - B;
+	glm::vec3 CD = D - C;
+	glm::vec3 DO = -D;
 
-	glm::dvec3 normalABD = normalize(cross(AD, BD));
+	glm::vec3 normalABD = normalize(cross(AD, BD));
 	if (dot(normalABD, -CD) > 0.0f)
 	{
 		normalABD *= -1.0f;
@@ -88,7 +88,7 @@ bool containsOriginTetrahedron(std::vector<glm::dvec3>& simplex){
 		simplex.erase(simplex.begin() + 2);
 		return false;
 	}
-	glm::dvec3 normalBCD = normalize(cross(BD, CD));
+	glm::vec3 normalBCD = normalize(cross(BD, CD));
 	if (dot(normalBCD, -AD) > 0.0f)
 	{
 		normalBCD *= -1.0f;
@@ -97,7 +97,7 @@ bool containsOriginTetrahedron(std::vector<glm::dvec3>& simplex){
 		simplex.erase(simplex.begin());
 		return false;
 	}
-	glm::dvec3 normalACD = normalize(cross(CD, AD));
+	glm::vec3 normalACD = normalize(cross(CD, AD));
 	if (dot(normalACD, -BD) > 0.0f)
 	{
 		normalACD *= -1.0f;
@@ -109,11 +109,11 @@ bool containsOriginTetrahedron(std::vector<glm::dvec3>& simplex){
 	return true;
 }
 
-bool isCollidingGJK(const std::vector<glm::dvec3>& shape1, const std::vector<glm::dvec3>& shape2,
+bool isCollidingGJK(const std::vector<glm::vec3>& shape1, const std::vector<glm::vec3>& shape2,
 	int maxIterations) {
 
-	std::vector<glm::dvec3> simplex;
-	glm::dvec3 direction = glm::dvec3(0.0f, 0.0f, 1.0f);
+	std::vector<glm::vec3> simplex;
+	glm::vec3 direction = glm::vec3(0.0f, 0.0f, 1.0f);
 	simplex.push_back(minkowskiPoint(shape1, shape2, direction));
 
 	if (!createLine(simplex, shape1, shape2)){
